+++ src/my_first_adventure_game/engine/graphics/__init__.py
@@ -1,3 +1,4 @@
+from my_first_adventure_game.engine.graphics.animation import Animation
 from my_first_adventure_game.engine.graphics.drawing import draw_text
 
-__all__ = ["draw_text"]
+__all__ = ["Animation", "draw_text"]